perf(wrap):`wrap`的析构函数入参新增存储对象的析构函数 style(lexer):移动`lexer`模块到单独的文件 feat(parser):在原`parser`模块中实现了基于递归下降四则运算的抽象语法树 feat(esp32):新增esp32 demo chore(*):优化构建结构 doc(readme):更新参考文献 chore(*):其他微小的工作

+++ src/common/candy_types.h
@@ -67,7 +67,7 @@ typedef struct candy_object * candy_object_t;
 typedef uint32_t        candy_hash_t;
 typedef candy_view_t    candy_string_t;
 typedef int64_t         candy_integer_t;
-typedef float           candy_float_t;
+typedef double          candy_float_t;
 typedef uint8_t         candy_boolean_t;
 typedef int (*candy_method_t)(candy_object_t);
